Extract: matcherPrototype to own file
