+++ lib/create-matcher.js
@@ -14,6 +14,7 @@ var valueToString = require("@sinonjs/commons").valueToString;
 var assertMethodExists = require("./create-matcher/assert-method-exists");
 var assertType = require("./create-matcher/assert-type");
 var isIterable = require("./create-matcher/is-iterable");
+var matcherPrototype = require("./create-matcher/matcher-prototype");
 
 var arrayIndexOf = arrayProto.indexOf;
 var arrayEvery = arrayProto.every;
@@ -27,12 +28,6 @@ var objectToString = objectProto.toString;
 
 var stringIndexOf = stringProto.indexOf;
 
-var matcher = {
-    toString: function() {
-        return this.message;
-    }
-};
-
 /**
  * Returns `true` when `object` is a matcher
  *
@@ -41,7 +36,7 @@ var matcher = {
  * @returns {boolean} Returns `true` when `object` is a matcher
  */
 function isMatcher(object) {
-    return isPrototypeOf(matcher, object);
+    return isPrototypeOf(matcherPrototype, object);
 }
 
 /**
@@ -147,7 +142,7 @@ var TYPE_MAP = {
  * @returns {object} A matcher object
  */
 function createMatcher(expectation, message) {
-    var m = Object.create(matcher);
+    var m = Object.create(matcherPrototype);
     var type = typeOf(expectation);
 
     if (message !== undefined && typeof message !== "string") {
@@ -174,40 +169,6 @@ function createMatcher(expectation, message) {
 
     return m;
 }
-
-matcher.or = function(valueOrMatcher) {
-    if (!arguments.length) {
-        throw new TypeError("Matcher expected");
-    }
-
-    var m2 = isMatcher(valueOrMatcher)
-        ? valueOrMatcher
-        : createMatcher(valueOrMatcher);
-    var m1 = this;
-    var or = Object.create(matcher);
-    or.test = function(actual) {
-        return m1.test(actual) || m2.test(actual);
-    };
-    or.message = m1.message + ".or(" + m2.message + ")";
-    return or;
-};
-
-matcher.and = function(valueOrMatcher) {
-    if (!arguments.length) {
-        throw new TypeError("Matcher expected");
-    }
-
-    var m2 = isMatcher(valueOrMatcher)
-        ? valueOrMatcher
-        : createMatcher(valueOrMatcher);
-    var m1 = this;
-    var and = Object.create(matcher);
-    and.test = function(actual) {
-        return m1.test(actual) && m2.test(actual);
-    };
-    and.message = m1.message + ".and(" + m2.message + ")";
-    return and;
-};
 
 createMatcher.isMatcher = isMatcher;
